gr-filter: gr-blocks: Add cpp flag
In commit 0cb3f2a, cpp support was added to the gr-filter yaml files. However, the cpp flag was not added to the low-pass and high-pass filter taps blocks.
This commit adds it in.

In commit 612c650, cpp support was added to the multiply matrix and unpacked to packed blocks, but the flags were not added. This commit adds them in.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>